# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: WMAX API
  description: Remote Patient Monitoring — production-grade clinical signal backend
  version: 2.0.0
servers:
- url: /
  description: Standart origin (Lokal dev :8000 yoki reverse proxy ildizi)
paths:
...
  /api/v1/patients:
    get:
      tags:
      - patients
      summary: 'Worklist sorted by acuity: red > no_data > amber > green.'
      operationId: list_patients_api_v1_patients_get
      security:
      - HTTPBearer: []
      parameters:
      - name: district
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by district
          title: District
        description: Filter by district
      - name: level
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - green
            - amber
            - red
            - no_data
            type: string
          - type: 'null'
          description: Filter by alert level
          title: Level
        description: Filter by alert level
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PatientSummary'
                title: Response List Patients Api V1 Patients Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - patients
      summary: Create a new patient
      operationId: create_patient_api_v1_patients_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Create Patient Api V1 Patients Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}:
    get:
      tags:
      - patients
      summary: Patient detail with multi-param series, baselines, trend, and prognosis.
      operationId: get_patient_detail_api_v1_patients__id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: days
        in: query
        required: false
        schema:
          type: integer
          maximum: 30
          minimum: 1
          description: History window in days
          default: 7
          title: Days
        description: History window in days
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientDetail'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    patch:
      tags:
      - patients
      summary: Patch patient profile
      operationId: patch_patient_api_v1_patients__id__patch
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientPatchRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Patch Patient Api V1 Patients  Id  Patch
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}/timeline:
    get:
      tags:
      - patients
      summary: Patient event timeline
      operationId: patient_timeline_api_v1_patients__id__timeline_get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                title: Response Patient Timeline Api V1 Patients  Id  Timeline Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}/readings:
    get:
      tags:
      - patients
      summary: Patient readings
      operationId: patient_readings_api_v1_patients__id__readings_get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                title: Response Patient Readings Api V1 Patients  Id  Readings Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}/baseline:
    get:
      tags:
      - patients
      summary: Current patient baseline
      operationId: patient_baseline_api_v1_patients__id__baseline_get
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                title: Response Patient Baseline Api V1 Patients  Id  Baseline Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}/baseline/approve:
    post:
      tags:
      - patients
      summary: Doctor approves personal baseline
      operationId: approve_baseline_contract_api_v1_patients__id__baseline_approve_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Approve Baseline Contract Api V1 Patients  Id  Baseline
                  Approve Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}/deceased:
    post:
      tags:
      - patients
      summary: Mark patient deceased
      operationId: mark_deceased_api_v1_patients__id__deceased_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Mark Deceased Api V1 Patients  Id  Deceased Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}/discharge:
    post:
      tags:
      - patients
      summary: Discharge patient and create 24h active-call follow-up task.
      operationId: discharge_patient_api_v1_patients__id__discharge_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}/approve-baseline:
    post:
      tags:
      - patients
      summary: Doctor approves personal baseline — transitions patient from learning
        to full phase.
      operationId: approve_baseline_api_v1_patients__id__approve_baseline_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string
                title: Response Approve Baseline Api V1 Patients  Id  Approve Baseline
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}/memberships:
    post:
      tags:
      - patients
      summary: Admit patient to a clinic
      operationId: create_membership_api_v1_patients__id__memberships_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipCreateRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Create Membership Api V1 Patients  Id  Memberships
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}/memberships/{mid}:
    delete:
      tags:
      - patients
      summary: Revoke patient clinic membership
      operationId: revoke_membership_api_v1_patients__id__memberships__mid__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: mid
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Mid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Revoke Membership Api V1 Patients  Id  Memberships  Mid  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}/memberships/{mid}/baseline-decision:
    post:
      tags:
      - patients
      summary: Record baseline ownership decision
      operationId: membership_baseline_decision_api_v1_patients__id__memberships__mid__baseline_decision_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: mid
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Mid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BaselineDecisionRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Membership Baseline Decision Api V1 Patients  Id  Memberships  Mid  Baseline
                  Decision Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}/access/invite:
    post:
      tags:
      - patients
      summary: Invite caregiver access
      operationId: invite_access_api_v1_patients__id__access_invite_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AccessInviteRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Invite Access Api V1 Patients  Id  Access Invite Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/patients/{id}/access/{aid}:
    delete:
      tags:
      - patients
      summary: Revoke caregiver access
      operationId: revoke_access_api_v1_patients__id__access__aid__delete
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: aid
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Aid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Revoke Access Api V1 Patients  Id  Access  Aid  Delete
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v1/patients/{id}/consents/{cid}/revoke:
    post:
      tags:
      - patients
      summary: Revoke patient consent
      operationId: revoke_consent_api_v1_patients__id__consents__cid__revoke_post
      security:
      - HTTPBearer: []
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Id
      - name: cid
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Cid
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                title: Response Revoke Consent Api V1 Patients  Id  Consents  Cid  Revoke
                  Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    BaselineDecisionRequest:
      properties:
        decision:
          type: string
          title: Decision
      type: object
      required:
      - decision
      title: BaselineDecisionRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MembershipCreateRequest:
      properties:
        tenant_id:
          type: string
          format: uuid
          title: Tenant Id
        kind:
          type: string
          const: care
          title: Kind
          default: care
      type: object
      required:
      - tenant_id
      title: MembershipCreateRequest
...
    PatientSummary:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        full_name:
          type: string
          title: Full Name
        age:
          type: integer
          title: Age
        sex:
          anyOf:
          - type: string
            enum:
            - m
            - f
          - type: 'null'
          title: Sex
        diagnosis:
          type: string
          title: Diagnosis
        district:
          type: string
          title: District
        phase:
          type: string
          enum:
          - calib
          - learning
          - full
          title: Phase
        level:
          type: string
          enum:
          - green
          - amber
          - red
          - no_data
          title: Level
        composite_score:
          type: number
          title: Composite Score
        triggered_params:
          additionalProperties:
            type: number
          type: object
          title: Triggered Params
          default: {}
        trend:
          $ref: '#/components/schemas/Trend'
        last_reading_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Last Reading At
        open_task:
          anyOf:
          - $ref: '#/components/schemas/Task'
          - type: 'null'
      type: object
      required:
      - id
      - full_name
      - age
      - diagnosis
      - district
      - phase
      - level
      - composite_score
      - trend
      title: PatientSummary
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
...